requirements.txt
Flask
Pillow
Werkzeug
pytz
twilio
watchdog>=4.0.0
python-dotenv